+++ src/Plutarch/Extra/Bool.hs
@@ -4,7 +4,7 @@ module Plutarch.Extra.Bool (
 
 import Data.Kind (Type)
 import Plutarch (S, Term)
-import Plutarch.Bool (POrd ((#<)), pif, (#==))
+import Plutarch.Bool (POrd, pif, (#<), (#==))
 
 {- | Perform a \'three-way\' comparison on two 'Term's, then return a result
  based on the outcome.