+++ src/Plutarch/Extra/Field.hs
@@ -6,6 +6,7 @@
 
 module Plutarch.Extra.Field (pletAll, pletAllC) where
 
+import Data.Kind (Type)
 import GHC.TypeLits (Symbol)
 import Plutarch.DataRepr.Internal.Field (
     BindFields,
@@ -16,7 +17,7 @@ import Plutarch.DataRepr.Internal.Field (
     PDataFields (PFields),
     pletFields,
  )
-import Plutarch.Prelude (PLabeledType (..), Term, TermCont, tcont)
+import Plutarch.Prelude (PLabeledType ((:=)), S, Term, TermCont, tcont)
 
 type family BindAll (ps :: [PLabeledType]) :: [Symbol] where
     BindAll '[] = '[]
@@ -24,10 +25,10 @@ type family BindAll (ps :: [PLabeledType]) :: [Symbol] where
 
 {- | Same as `pletFields` but instead of specifiying fields, it will take all fields.
 
- @since 1.2.1
+ @since 1.3.0
 -}
 pletAll ::
-    forall a s b ps bs.
+    forall (a :: S -> Type) (s :: S) (b :: S -> Type) (ps :: [PLabeledType]) bs.
     ( PDataFields a
     , ps ~ PFields a
     , bs ~ Bindings ps (BindAll ps)
@@ -40,10 +41,10 @@ pletAll = pletFields @(BindAll ps)
 
 {- | TermCont version of `pletAll`
 
- @since 1.2.1
+ @since 1.3.0
 -}
 pletAllC ::
-    forall a s b ps bs.
+    forall (a :: S -> Type) (s :: S) (b :: S -> Type) (ps :: [PLabeledType]) bs.
     ( PDataFields a
     , ps ~ PFields a
     , bs ~ Bindings ps (BindAll ps)