+++ src/Plutarch/Extra/Identity.hs
@@ -1,5 +1,5 @@
 {-# LANGUAGE DeriveAnyClass #-}
-{-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
 
@@ -8,11 +8,13 @@ module Plutarch.Extra.Identity (
 ) where
 
 import Data.Kind (Type)
-import Generics.SOP (I (I), Top)
-import Generics.SOP.TH (deriveGeneric)
+import GHC.Generics (Generic)
+import Generics.SOP (Top)
+import qualified Generics.SOP as SOP
 import Plutarch (
-    DerivePNewtype (DerivePNewtype),
+    DerivePlutusType (..),
     PlutusType,
+    PlutusTypeNewtype,
     S,
     Term,
     pcon,
@@ -21,7 +23,7 @@ import Plutarch (
     unTermCont,
     (#),
  )
-import Plutarch.Bool (PEq, POrd)
+import Plutarch.Bool (PEq, POrd, PPartialOrd)
 import Plutarch.Builtin (PIsData)
 import Plutarch.Extra.Applicative (PApplicative (ppure), PApply (pliftA2))
 import Plutarch.Extra.Boring (PBoring (pboring))
@@ -32,6 +34,7 @@ import Plutarch.Extra.Comonad (
 import Plutarch.Extra.Functor (PFunctor (PSubcategory, pfmap))
 import Plutarch.Extra.TermCont (pmatchC)
 import Plutarch.Integer (PIntegral)
+import Plutarch.Num (PNum)
 import Plutarch.Show (PShow)
 
 {- | Just a value of type @a@.
@@ -40,53 +43,38 @@ import Plutarch.Show (PShow)
 -}
 newtype PIdentity (a :: S -> Type) (s :: S)
     = PIdentity (Term s a)
+    deriving stock
+        ( -- | @since 1.0.0
+          Generic
+        )
+    deriving anyclass
+        ( -- | @since 1.0.0
+          SOP.Generic
+        , -- | @since 1.0.0
+          PlutusType
+        )
 
-deriveGeneric ''PIdentity
-
--- | @since 1.0.0
-deriving via (DerivePNewtype (PIdentity a) a) instance (PlutusType (PIdentity a))
-
--- | @since 1.0.0
-deriving via
-    (DerivePNewtype (PIdentity a) a)
-    instance
-        (PIsData a) => (PIsData (PIdentity a))
-
--- | @since 1.0.0
-deriving via
-    (DerivePNewtype (PIdentity a) a)
-    instance
-        (PEq a) => PEq (PIdentity a)
+-- | @since 1.4.0
+instance DerivePlutusType (PIdentity a) where
+    type DPTStrat _ = PlutusTypeNewtype
 
 -- | @since 1.0.0
-deriving via
-    (DerivePNewtype (PIdentity a) a)
-    instance
-        (POrd a) => POrd (PIdentity a)
+deriving anyclass instance (PIsData a) => (PIsData (PIdentity a))
 
 -- | @since 1.0.0
-deriving via
-    (DerivePNewtype (PIdentity a) a)
-    instance
-        (PIntegral a) => PIntegral (PIdentity a)
+deriving anyclass instance (PEq a) => PEq (PIdentity a)
+
+-- | @since 1.4.0
+deriving anyclass instance (POrd a) => PPartialOrd (PIdentity a)
 
 -- | @since 1.0.0
-deriving via
-    (Term s (DerivePNewtype (PIdentity a) a))
-    instance
-        (Num (Term s a)) => Num (Term s (PIdentity a))
+deriving anyclass instance (POrd a) => POrd (PIdentity a)
 
 -- | @since 1.0.0
-deriving via
-    (Term s (DerivePNewtype (PIdentity a) a))
-    instance
-        (Semigroup (Term s a)) => Semigroup (Term s (PIdentity a))
+deriving anyclass instance (PIntegral a) => PIntegral (PIdentity a)
 
--- | @since 1.0.0
-deriving via
-    (Term s (DerivePNewtype (PIdentity a) a))
-    instance
-        (Monoid (Term s a)) => Monoid (Term s (PIdentity a))
+-- | @since 1.4.0
+deriving anyclass instance (PNum a) => PNum (PIdentity a)
 
 -- | @since 1.0.0
 deriving anyclass instance (PShow a) => PShow (PIdentity a)