+++ test/Spec/Api/V1/Value.hs
@@ -1,9 +1,0 @@
-module Spec.Api.V1.Value (tests) where
-
-import qualified Spec.Api.V1.Value.Unsorted as Unsorted
-import Test.Tasty (TestTree, testGroup)
-
-tests :: [TestTree]
-tests =
-    [ testGroup "unsorted values" Unsorted.tests
-    ]