+++ CHANGELOG.md
@@ -2,6 +2,54 @@
 
 This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
+## 1.4.0 -- 2022-08-02
+
+### Added 
+ - A `Plutarch.Oprhans` module, holding downcasted instances of semigroup and monoid when the upcasted type has the appropriate instances.
+ - `pflip` to `Plutarch.Extra.Function`
+ - `Plutarch.Extra.IsData`:
+   - `PDerivePDataFieldsViaDataList`
+   - A `PlutusTypeEnumData` as a deriving strategy for `PlutusType` 
+
+### Changed
+
+ - Update to [`Liqwid.nix`](https://github.com/liqwid-Labs/liqwid-nix)
+ - Update to Plutarch version 1.2. See the [CHANGELOG](https://github.com/Plutonomicon/plutarch-plutus/blob/v1.2.0/CHANGELOG.md) 
+   for full details.
+   - The flake now points at the `Plutonomicon` repository, instead of the Liqwid Labs fork.
+   - Changes to deriving strategies and constraints may cause some API breakage. In particular,
+     `deriving via`, `PMatch`, `PCon` has been eliminated, and redundant `PAsDAta`, `pfromData` have been reduced.
+
+### Removed
+
+ - The `Plutarch.Extra.Other` module has been removed. This held `deriving via` wrappers that are no longer necessary.
+ - Tests relating to `Value`s and unsorted `Map`s, since `Plutarch 1.2` removed the `PEq` constraint on unsorted maps.
+
+## 1.3.0 -- 2022-07-20
+
+### Added
+
+- `pmatchAll` and `pmatchAllC`, `pletFields` that gets all Plutarch record fields.
+- `Plutarch.Extra.MultiSig`, a basic N of M multisignature validation function.
+- `pscriptHashFromAddress`, gets script hash from an address.
+- `pisScriptAddress`, checks if given address is script address.
+- `pisPubKey`, checks if given credential is a pubkey hash.
+- `pfindOutputsToAddress`, finds all TxOuts sent to an Address.
+- `pfindTxOutDatum`, finds the data corresponding to a TxOut, if there is one.
+- `phasOnlyOneTokenOfCurrencySymbol`, checks if entire value only contain one token of given currency symbol.
+- `pon`, mirroring `Data.Function.on`.
+- `pbuiltinUncurry`, mirroring `uncurry`.
+- `pmaybeData`, mirroring `maybe` for `PMaybeData`.
+- `pdjust` for easier construction of `PDJust` value.
+- `pdnothing` for easier construction `PDNothing` value.
+
+### Modified
+
+- Fixed `PApplicative` instances that previously not worked due to not using `pfix`.
+- Renamed `PType` to `S -> Type`.
+- Renamed `mustBePJust` to `passertPJust`.
+- Renamed `mustBePDJust` to `passertPDJust`.
+
 ## 1.2.0 -- 2022-07-12
 
 ### Added
@@ -34,6 +82,11 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 ## 1.1.0 -- 2022-06-17
 
 ### Added
+
+- Convenience wrapper for `DerivePNewtype`: `DerivePNewtype'`, `DerivePConstantViaNewtype'`
+- Encode product types as lists: `ProductIsData`, `DerivePConstantViaDataList`
+- Encode enum types as integers: `EnumIsData`, `PEnumData` and `DerivePConstantViaEnum`
+- Plutarch helper functions: `pmatchEnum`, `pmatchEnumFromData`
 
 #### AssocMap (`Plutarch.Extra.Map`)
 
@@ -102,6 +155,10 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 - `ptryFindDatum`
 - `pfindDatum`
 - `pfindTxInByTxOutRef`
+
+### Modified
+
+- Rename `PConstantViaDataList` to `DerivePConstantViaDataList`
 
 ## 1.0.0 -- 2022-05-24